src/pyproject.toml
[tool.poetry]
name = "ritengine"
version = "0.1.0"
description = ""
authors = ["Mazdak <me@mazdak.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.10.5"
aiohttp-retry = "^2.8.3"
aiosignal = "^1.3.1"
amqp = "^5.2.0"
annotated-types = "^0.7.0"
anyio = "^4.4.0"
asgiref = "^3.8.1"
async-timeout = "^4.0.3"
attrs = "^24.2.0"
autobahn = "^24.4.2"
automat = "^24.8.1"
billiard = "^4.2.0"
celery = "^5.4.0"
certifi = "^2024.8.30"
cffi = "^1.17.1"
channels = "^4.1.0"
charset-normalizer = "^3.3.2"
click = "^8.1.7"
click-didyoumean = "^0.3.1"
click-plugins = "^1.1.1"
click-repl = "^0.3.0"
colorama = "^0.4.6"
constantly = "^23.10.4"
cron-descriptor = "^1.4.5"
cryptography = "^43.0.1"
daphne = "^4.1.2"
distro = "^1.9.0"
dj-database-url = "^2.2.0"
dj-rest-auth = "^6.0.0"
django = "^5.1.1"
django-admin-tools = "^0.9.3"
django-allauth = "^64.2.1"
django-celery-beat = "^2.7.0"
django-cors-headers = "^4.4.0"
django-filter = "^24.3"
django-otp = "^1.5.4"
django-phonenumber-field = "^8.0.0"
django-redis = "^5.4.0"
django-timezone-field = "^7.0"
django-unfold = "^0.38.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
exceptiongroup = "^1.2.2"
faker = "^28.4.1"
flower = "^2.0.1"
frozenlist = "^1.4.1"
h11 = "^0.14.0"
hiredis = "^3.0.0"
httpcore = "^1.0.5"
httpx = "^0.27.2"
humanize = "^4.10.0"
hyperlink = "^21.0.0"
idna = "^3.8"
incremental = "^24.7.2"
inflection = "^0.5.1"
jsonschema = "^4.23.0"
jsonschema-specifications = "^2023.12.1"
kombu = "^5.4.0"
multidict = "^6.1.0"
oauthlib = "^3.2.2"
openai = "^1.44.1"
packaging = "^24.1"
phonenumbers = "^8.13.45"
pillow = "^10.4.0"
prometheus-client = "^0.20.0"
prompt-toolkit = "^3.0.47"
psycopg2 = "^2.9.9"
pyasn1 = "^0.6.0"
pyasn1-modules = "^0.4.0"
pycparser = "^2.22"
pydantic = "^2.9.1"
pydantic-core = "^2.23.3"
pyjwt = "^2.9.0"
pyopenssl = "^24.2.1"
pyotp = "^2.9.0"
python-crontab = "^3.2.0"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
pyyaml = "^6.0.2"
redis = "^5.0.8"
referencing = "^0.35.1"
requests = "^2.32.3"
requests-oauthlib = "^2.0.0"
rpds-py = "^0.20.0"
service-identity = "^24.1.0"
six = "^1.16.0"
sniffio = "^1.3.1"
sqlparse = "^0.5.1"
tenacity = "^9.0.0"
tomli = "^2.0.1"
tornado = "^6.4.1"
tqdm = "^4.66.5"
twilio = "^9.3.0"
twisted = "^24.7.0"
txaio = "^23.1.1"
typing-extensions = "^4.12.2"
tzdata = "^2024.1"
uritemplate = "^4.1.1"
urllib3 = "^2.2.2"
vine = "^5.1.0"
wcwidth = "^0.2.13"
yarl = "^1.11.1"
zope-interface = "^7.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
